+++ test/components/datagrid/datagrid.puppeteer-spec.js
@@ -56,6 +56,32 @@ describe('Datagrid', () => {
     });
   });
 
+  describe('Inline Editor', () => {
+    const url = `${baseUrl}/test-editable-with-inline-editor.html`;
+
+    beforeAll(async () => {
+      await page.goto(url, { waitUntil: ['domcontentloaded', 'networkidle0'] });
+    });
+
+    it('Should have inline editors', async () => {
+      await page.waitForSelector('.has-inline-editor', { visible: true })
+        .then(element => expect(element).toBeTruthy());
+    });
+
+    it('Should be able to clean editors on click and backspace', async () => {
+      await page.waitForSelector('.has-inline-editor', { visible: true })
+        .then(element => expect(element).toBeTruthy());
+
+      await page.waitForSelector('#datagrid-inline-input-1-2', { visible: true })
+        .then(element => expect(element).toBeTruthy());
+
+      await page.click('#datagrid-inline-input-1-2');
+
+      await page.keyboard.press('Backspace');
+      expect(await page.$eval('#datagrid-inline-input-1-2', el => el.value)).toBe('');
+    });
+  });
+
   describe('Tree row status', () => {
     const url = `${baseUrl}/test-tree-rowstatus.html`;
     beforeAll(async () => {
@@ -419,6 +445,33 @@ describe('Datagrid', () => {
       await priceCell.click();
       await page.keyboard.press('Enter');
       await priceCell.evaluate(el => el.textContent).then(text => expect(text).toEqual('1,500'));
+    });
+  });
+
+  describe('Extra class for tooltip tests', () => {
+    const url = `${baseUrl}/example-header-icon-with-tooltip`;
+
+    beforeAll(async () => {
+      await page.goto(url, { waitUntil: ['domcontentloaded', 'networkidle0'] });
+    });
+
+    it('should show tooltip when hovered', async () => {
+      const th = 'example-header-icon-with-tooltip-datagrid-1-header-2';
+      await page.hover(`#${th}`);
+
+      await page.waitForSelector('#example-header-icon-with-tooltip-datagrid-0tooltip', { visible: true })
+        .then(element => element.getProperty('className'))
+        .then(className => className.jsonValue())
+        .then(classNameString => expect(classNameString).not.toContain('is-hidden'));
+    });
+
+    it('should show tooltip when hovered', async () => {
+      await page.hover('.icon.datagrid-header-icon');
+
+      await page.waitForSelector('#example-header-icon-with-tooltip-datagrid-0tooltip', { visible: true })
+        .then(element => element.getProperty('className'))
+        .then(className => className.jsonValue())
+        .then(classNameString => expect(classNameString).toContain('tooltip-extra-class'));
     });
   });
 });